Insertions now working for oxdna system. Code cleaning

+++ Assets/Scripts/files/OxViewFile.cs
@@ -33,9 +33,9 @@ public class OxViewMonomer
     public int Id { get; set; }
     public string Type { get; set; } = "";
     public string Class { get; set; } = "";
-    public List<float> P { get; set; } = new List<float>();
-    public List<float> A1 { get; set; } = new List<float>();
-    public List<float> A3 { get; set; } = new List<float>();
+    public List<double> P { get; set; } = new List<double>();
+    public List<double> A1 { get; set; } = new List<double>();
+    public List<double> A3 { get; set; } = new List<double>();
     public int? N3 { get; set; }
     public int? N5 { get; set; }
     public int? Bp { get; set; }
